Add unit tests for cmd/build/download.go and merge download artifacts logic for both bk build download and bk artifacts download
